Refactor authentication schema and middleware
- Removed the legacy `auth-schema.ts` file and migrated the user, session, account, and verification schemas to `lib/db/schema.ts`.
- Updated middleware to utilize `NextResponse` for redirects instead of the deprecated `Response` object, improving compatibility with Next.js.
- Simplified session handling in the `SessionProvider` component by streamlining the session conversion logic.

These changes enhance the authentication structure and align with current Next.js practices.
